+++ changelog.md
@@ -8,6 +8,7 @@
 - Reducing the left and right Padding on LSX Buttons from 40px to 15px.
 - A `lsx_blocks_latest_posts_carousel_meta` to allow you to filter the meta.
 - Adding compatibility with WordPress 5.5.
+- Added color options for the Team block.
 
 ### Fixed